fix: Fix package name of SpaceX API Service
