+++ services/spacex/src/main/kotlin/com/haroldadmin/moonshot/services/spacex/v4/GenericModels.kt
@@ -1,0 +1,3 @@
+package com.haroldadmin.moonshot.services.spacex.v4
+
+import com.squareup.moshi.JsonClass
