--- services/spacex/src/main/kotlin/com/haroldadmin/services/spacex/v4/GenericModels.kt
@@ -1,3 +1,0 @@
-package com.haroldadmin.services.spacex.v4
-
-import com.squareup.moshi.JsonClass
